pyproject.toml
[project]
name = "lahacks2025"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "faiss-cpu>=1.11.0",
    "numpy>=2.3.1",
    "olefile>=0.47",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.1",
    "pyside6>=6.9.1",
    "pystray>=0.19.5",
    "python-docx>=1.2.0",
    "qdrant-client>=1.14.3",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "watchdog>=6.0.0",
]

[project.scripts]
mckn = "munchkin:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
